requirements.txt
python-dotenv
openai
supabase
markdown
PyGithub
